working on windows paralellization, added trimOligos class to be used by trim.flows and trim.seqs commands.  Added sequenceParser class to divide fasta and names files by group.  Added group and bygroup options to pre.cluster to allow you to pre.cluster within each group.

+++ alignmentdb.cpp
@@ -77,7 +77,7 @@ AlignmentDB::AlignmentDB(string fastaFileName, string s, int kmerSize, float gap
 			
 			#ifdef USE_MPI	
 				int pid, processors;
-				vector<unsigned long int> positions;
+				vector<unsigned long long> positions;
 			
 				MPI_Status status; 
 				MPI_File inMPI;
