+++ chimera.h
@@ -133,9 +133,9 @@ struct sim {
 };
 
 struct linePair {
-			unsigned long int start;
-			unsigned long int end;
-			linePair(unsigned long int i, unsigned long int j) : start(i), end(j) {}
+			unsigned long long start;
+			unsigned long long end;
+			linePair(unsigned long long i, unsigned long long j) : start(i), end(j) {}
 			linePair(){}
 };
